#1855 Modified project structure for testing purposes. Moved test entities from `src/test/resources` to `/src/main/java`, so it's no longer necessary to have that additional directory on your build path. Enabled annotation processing during the test-compile build phase in order to generate meta-models for test entities. Generated sources are put under `target/generated-test-sources/test-annotations`. It would be nice to get rid of the `test-annotations` part, but maven doesn't comply, which is not a big deal. Also meta-models are now generated only for top-level classes, since nested ones are not supported yet.
