+++ platform-annotation-processors/src/test/java/ua/com/fielden/platform/processors/test_entities/TestEntityNotPersistent.java
@@ -1,6 +1,4 @@
-package ua.com.fielden.platform.processors.metamodel.test_entities;
-
-import java.util.Date;
+package ua.com.fielden.platform.processors.test_entities;
 
 import ua.com.fielden.platform.annotations.metamodel.DomainEntity;
 import ua.com.fielden.platform.entity.AbstractEntity;
@@ -8,7 +6,6 @@ import ua.com.fielden.platform.entity.annotation.DescTitle;
 import ua.com.fielden.platform.entity.annotation.IsProperty;
 import ua.com.fielden.platform.entity.annotation.KeyTitle;
 import ua.com.fielden.platform.entity.annotation.KeyType;
-import ua.com.fielden.platform.entity.annotation.MapEntityTo;
 import ua.com.fielden.platform.entity.annotation.MapTo;
 import ua.com.fielden.platform.entity.annotation.Observable;
 import ua.com.fielden.platform.entity.annotation.Title;
@@ -23,42 +20,26 @@ import ua.com.fielden.platform.utils.Pair;
  */
 @KeyType(String.class)
 @KeyTitle("Key")
-@MapEntityTo
 @DomainEntity
 @DescTitle("Description")
-public class TestEntityParent extends AbstractEntity<String> {
+public class TestEntityNotPersistent extends AbstractEntity<String> {
 
-    private static final Pair<String, String> entityTitleAndDesc = TitlesDescsGetter.getEntityTitleAndDesc(TestEntityParent.class);
+    private static final Pair<String, String> entityTitleAndDesc = TitlesDescsGetter.getEntityTitleAndDesc(TestEntityNotPersistent.class);
     public static final String ENTITY_TITLE = entityTitleAndDesc.getKey();
     public static final String ENTITY_DESC = entityTitleAndDesc.getValue();
     
     @IsProperty
     @MapTo
     @Title(value = "Prop1")
-    private String prop1;
-    
-    @IsProperty
-    @MapTo
-    @Title(value = "Prop2", desc = "Extended_description")
-    private Date prop2;
-
-    @Observable
-    public TestEntityParent setProp2(final Date prop2) {
-        this.prop2 = prop2;
-        return this;
-    }
-
-    public Date getProp2() {
-        return prop2;
-    }
+    private int prop1;
     
     @Observable
-    public TestEntityParent setProp1(final String prop1) {
+    public TestEntityNotPersistent setProp1(final int prop1) {
         this.prop1 = prop1;
         return this;
     }
 
-    public String getName() {
+    public int getProp1() {
         return prop1;
     }
 }
