+++ platform-annotation-processors/src/test/java/ua/com/fielden/platform/processors/test_entities/TestEntitySinkNodesOnly.java
@@ -1,6 +1,9 @@
-package ua.com.fielden.platform.processors.metamodel.test_entities;
+package ua.com.fielden.platform.processors.test_entities;
+
+import java.util.Date;
 
 import ua.com.fielden.platform.annotations.metamodel.DomainEntity;
+import ua.com.fielden.platform.entity.AbstractEntity;
 import ua.com.fielden.platform.entity.annotation.DescRequired;
 import ua.com.fielden.platform.entity.annotation.DescTitle;
 import ua.com.fielden.platform.entity.annotation.DisplayDescription;
@@ -25,39 +28,41 @@ import ua.com.fielden.platform.utils.Pair;
 @MapEntityTo
 @DomainEntity
 @DescTitle("Description")
-public class TestEntityChild extends TestEntityParent {
+@DisplayDescription
+@DescRequired
+public class TestEntitySinkNodesOnly extends AbstractEntity<String> {
 
-    private static final Pair<String, String> entityTitleAndDesc = TitlesDescsGetter.getEntityTitleAndDesc(TestEntityChild.class);
+    private static final Pair<String, String> entityTitleAndDesc = TitlesDescsGetter.getEntityTitleAndDesc(TestEntitySinkNodesOnly.class);
     public static final String ENTITY_TITLE = entityTitleAndDesc.getKey();
     public static final String ENTITY_DESC = entityTitleAndDesc.getValue();
     
     @IsProperty
     @MapTo
     @Title(value = "Prop1")
-    private int prop1;
+    private String prop1;
     
     @IsProperty
     @MapTo
-    @Title(value = "Parent entity", desc = "Extended_description")
-    private TestEntityParent parent;
+    @Title(value = "Prop2")
+    private Date prop2;
 
     @Observable
-    public TestEntityChild setParent(final TestEntityParent parent) {
-        this.parent = parent;
-        return this;
-    }
-
-    public TestEntityParent getParent() {
-        return parent;
-    }
-    
-    @Observable
-    public TestEntityChild setProp1(final int prop1) {
+    public TestEntitySinkNodesOnly setProp1(final String prop1) {
         this.prop1 = prop1;
         return this;
     }
 
-    public int getProp1() {
+    public String getName() {
         return prop1;
+    }
+
+    @Observable
+    public TestEntitySinkNodesOnly setProp2(final Date prop2) {
+        this.prop2 = prop2;
+        return this;
+    }
+
+    public Date getProp2() {
+        return prop2;
     }
 }
