+++ platform-annotation-processors/src/test/java/ua/com/fielden/platform/processors/test_entities/TestEntityWithDescTitle.java
@@ -1,0 +1,33 @@
+package ua.com.fielden.platform.processors.test_entities;
+
+import ua.com.fielden.platform.annotations.metamodel.DomainEntity;
+import ua.com.fielden.platform.entity.AbstractEntity;
+import ua.com.fielden.platform.entity.annotation.DescRequired;
+import ua.com.fielden.platform.entity.annotation.DescTitle;
+import ua.com.fielden.platform.entity.annotation.DisplayDescription;
+import ua.com.fielden.platform.entity.annotation.KeyTitle;
+import ua.com.fielden.platform.entity.annotation.KeyType;
+import ua.com.fielden.platform.entity.annotation.MapEntityTo;
+import ua.com.fielden.platform.reflection.TitlesDescsGetter;
+import ua.com.fielden.platform.utils.Pair;
+
+/**
+ * Master entity object.
+ *
+ * @author Developers
+ *
+ */
+@KeyType(String.class)
+@KeyTitle("Key")
+@MapEntityTo
+@DomainEntity
+@DescTitle("Description")
+@DisplayDescription
+@DescRequired
+public class TestEntityWithDescTitle extends AbstractEntity<String> {
+
+    private static final Pair<String, String> entityTitleAndDesc = TitlesDescsGetter.getEntityTitleAndDesc(TestEntityWithDescTitle.class);
+    public static final String ENTITY_TITLE = entityTitleAndDesc.getKey();
+    public static final String ENTITY_DESC = entityTitleAndDesc.getValue();
+    
+}
